demos/bank/src/step2/aiozmq/requirements.txt
msgpack==1.1.1
aiozmq==1.0.0
